Added logout functionality to Navbar
