+++ src/components/NavBar.js
@@ -2,10 +2,26 @@ import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
 import NavDropdown from 'react-bootstrap/NavDropdown';
+import { auth } from '../pages/firebase';
+import { useNavigate } from 'react-router-dom';
+import {  signOut } from "firebase/auth";
 import '../design/Navbar.css';
 import { FaHome } from 'react-icons/fa';
 
 function ColorSchemesExample() {
+
+  const navigate = useNavigate(); 
+
+  const handleLogout = () => {               
+    signOut(auth).then(() => {
+    // Sign-out successful.
+        navigate("/login");
+        console.log("Signed out successfully")
+    }).catch((error) => {
+    // An error happened.
+    });
+}
+
   return (
       <Navbar bg="dark" variant="dark" >
         <Container>
@@ -21,7 +37,7 @@ function ColorSchemesExample() {
                     <NavDropdown.Item href="/">Privacy policy</NavDropdown.Item>
                     <NavDropdown.Item href="/">Contact Us</NavDropdown.Item>
                     <NavDropdown.Divider />
-                    <NavDropdown.Item href="/">
+                    <NavDropdown.Item href="/" onClick={handleLogout}>
                         Logout
                     </NavDropdown.Item>
                 </NavDropdown>
