+++ src/pages/Home.js
@@ -1,8 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { onAuthStateChanged } from "firebase/auth";
 import { auth } from './firebase';
-import { useNavigate } from 'react-router-dom';
-import {  signOut } from "firebase/auth";
 import Login from './Login';
 import NavBar from '../components/NavBar';
  
@@ -20,18 +18,6 @@ const Home = () => {
          
     }, [])
 
-    const navigate = useNavigate();
-
-    const handleLogout = () => {               
-        signOut(auth).then(() => {
-        // Sign-out successful.
-            navigate("/");
-            console.log("Signed out successfully")
-        }).catch((error) => {
-        // An error happened.
-        });
-    }
-
     return(
         <div>
             {!currentUser ? (
@@ -45,12 +31,6 @@ const Home = () => {
                 <p>
                     Welcome Home {currentUser.email}
                 </p>
-
-                <div>
-                    <button onClick={handleLogout}>
-                        Logout
-                    </button>
-                </div>
             </nav>)}
         </div>
     )
